# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Rocket Leaf Private Desktop API
  version: 2.0.0
  description: Private loopback API between the Electron main process and local rocket-leafd; not published as a public API.
servers:
  - url: http://127.0.0.1:{port}/v1
    variables:
      port: { default: '0' }
security:
  - bearerAuth: []
paths:
...
  /connections:
    get:
      operationId: listConnections
      responses:
        '200':
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { type: array, items: { $ref: '#/components/schemas/Connection' } } },
              },
          }
    post:
      operationId: addConnection
      requestBody:
        {
          required: true,
          content:
            { application/json: { schema: { $ref: '#/components/schemas/ConnectionInput' } } },
        }
      responses:
        {
          '201':
            {
              description: Created,
              content:
                { application/json: { schema: { $ref: '#/components/schemas/Connection' } } },
            },
        }
  /connections/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    put:
      operationId: updateConnection
      requestBody:
        {
          required: true,
          content:
            { application/json: { schema: { $ref: '#/components/schemas/ConnectionInput' } } },
        }
      responses:
        {
          '200':
            {
              description: OK,
              content:
                { application/json: { schema: { $ref: '#/components/schemas/Connection' } } },
            },
        }
    delete: { operationId: deleteConnection, responses: { '204': { description: Deleted } } }
  /connections/{id}/connect:
    post:
      {
        operationId: connect,
        parameters: [{ name: id, in: path, required: true, schema: { type: integer } }],
        responses: { '204': { description: Connected } },
      }
  /connections/{id}/disconnect:
    post:
      {
        operationId: disconnect,
        parameters: [{ name: id, in: path, required: true, schema: { type: integer } }],
        responses: { '204': { description: Disconnected } },
      }
  /connections/{id}/default:
    post:
      {
        operationId: setDefaultConnection,
        parameters: [{ name: id, in: path, required: true, schema: { type: integer } }],
        responses: { '204': { description: Updated } },
      }
  /connections/{id}/test:
    post:
      {
        operationId: testConnection,
        parameters: [{ name: id, in: path, required: true, schema: { type: integer } }],
        responses: { '200': { description: OK } },
      }
...
  /settings:
    get:
      {
        operationId: getSettings,
        responses:
          {
            '200':
              {
                description: OK,
                content:
                  { application/json: { schema: { $ref: '#/components/schemas/AppSettings' } } },
              },
          },
      }
    put:
      operationId: updateSettings
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: '#/components/schemas/AppSettings' } } },
        }
      responses:
        {
          '200':
            {
              description: OK,
              content:
                { application/json: { schema: { $ref: '#/components/schemas/AppSettings' } } },
            },
        }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
...
    ConnectionInput:
      type: object
      required: [name, env, nameServer, timeoutSec, enableACL, accessKey, secretKey, remark]
      properties:
        name: { type: string }
        env: { type: string, enum: [production, test, development] }
        nameServer: { type: string }
        timeoutSec: { type: integer }
        enableACL: { type: boolean }
        accessKey: { type: string }
        secretKey: { type: string }
        remark: { type: string }
        credentialsMode: { type: string, enum: [preserve, replace, clear] }
    Connection:
      allOf:
        - $ref: '#/components/schemas/ConnectionInput'
        - type: object
          required: [id, status, lastCheck, isDefault, accessKeyConfigured, secretKeyConfigured]
          properties:
            id: { type: integer }
            status: { type: string, enum: [online, offline] }
            lastCheck: { type: string }
            isDefault: { type: boolean }
            accessKeyConfigured: { type: boolean }
            secretKeyConfigured: { type: boolean }
    AppSettings:
      type: object
      required:
        [
          theme,
          language,
          fontSize,
          uiFont,
          monospaceFont,
          autoConnectLast,
          connectTimeoutMs,
          requestTimeoutMs,
          globalAccessKey,
          globalSecretKey,
          skipTlsVerify,
          proxyEnabled,
          proxyType,
          proxyHost,
          proxyPort,
          lagAlertThreshold,
          diskAlertThreshold,
          desktopNotifications,
          timezone,
          timestampFormat,
          autoFormatJson,
          maxPayloadRenderBytes,
          fetchLimit,
        ]
      properties:
        theme: { type: string }
        language: { type: string }
        fontSize: { type: integer }
        uiFont: { type: string }
        monospaceFont: { type: string }
        autoConnectLast: { type: boolean }
        connectTimeoutMs: { type: integer }
        requestTimeoutMs: { type: integer }
        globalAccessKey: { type: string }
        globalSecretKey: { type: string }
        globalAccessKeyConfigured: { type: boolean }
        globalSecretKeyConfigured: { type: boolean }
        globalCredentialsMode: { type: string, enum: [preserve, replace, clear] }
        skipTlsVerify: { type: boolean }
        proxyEnabled: { type: boolean }
        proxyType: { type: string }
        proxyHost: { type: string }
        proxyPort: { type: string }
        lagAlertThreshold: { type: integer }
        diskAlertThreshold: { type: integer }
        desktopNotifications: { type: boolean }
        timezone: { type: string }
        timestampFormat: { type: string }
        autoFormatJson: { type: boolean }
        maxPayloadRenderBytes: { type: integer }
        fetchLimit: { type: integer }
...